+++ sphinx/source/changelog.rst
@@ -7,6 +7,8 @@ Changelog
   in the memory snapshot)
 * Added the :ref:`WHILE` macro (for repeatedly expanding macros until a
   conditional expression becomes false)
+* Added the :ref:`UDGS` macro (as an alternative to the :ref:`UDGARRAY` macro
+  for creating an image of a rectangular array of UDGs)
 * Added support to the :ref:`DEF` macro for using replacement fields to
   represent the defined macro's argument values, and for stripping leading and
   trailing whitespace from the defined macro's output